fix: Publishing phase should run after git push (#33)
Fix #30

Before this change the `publish` phase was run *before* the `git push`. This can be problematic because if the `git push` fails then the software has been published, but the repository hasn't been updated.

This PR fixes it by establishing or more healthy order of the tasks. First update the git repository (commit, tag AND push). And only after succesful push: publish.

That way we publish only if the update of the git repository was successful.

One may ask what if the publish step fail? Well, given the `verify` phase was successful, we can consider the tagged commit has being publishable. If the publication fails, it means there was some unfortunate network issues (or similar), but nothing blocking. We expect the user to be able to manually checkout the tag and manually finish the publication process.